# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job Finder API",
        "description": "API para el sistema de búsqueda y gestión de empleos",
        "contact": {
            "name": "Soporte Job Finder",
            "email": "admin@jobfinder.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/api/jobs": {
            "get": {
                "tags": [
                    "Trabajos"
                ],
                "summary": "Obtener lista de trabajos",
                "description": "Obtiene una lista paginada de trabajos con filtros opcionales",
                "operationId": "getJobs",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Término de búsqueda",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "desarrollador"
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Filtrar por estado activo",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    },
                    {
                        "name": "min_salary",
                        "in": "query",
                        "description": "Salario mínimo",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "example": 2000000
                        }
                    },
                    {
                        "name": "max_salary",
                        "in": "query",
                        "description": "Salario máximo",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "example": 5000000
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Campo para ordenar",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "created_at"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trabajos obtenidos exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Trabajos obtenidos exitosamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jobs/{id}": {
            "get": {
                "tags": [
                    "Trabajos"
                ],
                "summary": "Obtener trabajo específico",
                "description": "Obtiene la información detallada de un trabajo",
                "operationId": "getJob",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del trabajo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trabajo obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Trabajo obtenido exitosamente"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Trabajo no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Trabajo no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/jobs": {
            "post": {
                "tags": [
                    "Administración"
                ],
                "summary": "Crear nuevo trabajo",
                "description": "Crea una nueva oferta de trabajo (solo administradores)",
                "operationId": "createJob",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "descripcion",
                                    "empresa",
                                    "ubicacion",
                                    "tipo_contrato",
                                    "salario_min",
                                    "salario_max"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Desarrollador Full Stack"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Buscamos un desarrollador con experiencia en Laravel y Vue.js"
                                    },
                                    "empresa": {
                                        "type": "string",
                                        "example": "TechCorp"
                                    },
                                    "ubicacion": {
                                        "type": "string",
                                        "example": "Bogotá, Colombia"
                                    },
                                    "tipo_contrato": {
                                        "type": "string",
                                        "enum": [
                                            "tiempo_completo",
                                            "tiempo_parcial",
                                            "contrato",
                                            "freelance"
                                        ],
                                        "example": "tiempo_completo"
                                    },
                                    "salario_min": {
                                        "type": "number",
                                        "example": 3000000
                                    },
                                    "salario_max": {
                                        "type": "number",
                                        "example": 5000000
                                    },
                                    "requisitos": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Laravel",
                                            "Vue.js",
                                            "MySQL"
                                        ]
                                    },
                                    "beneficios": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Seguro médico",
                                            "Bonos de productividad"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Trabajo creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Trabajo creado exitosamente"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/admin/jobs/stats": {
            "get": {
                "tags": [
                    "Administración"
                ],
                "summary": "Obtener estadísticas de trabajos",
                "description": "Obtiene estadísticas detalladas de los trabajos",
                "operationId": "getJobStats",
                "responses": {
                    "200": {
                        "description": "Estadísticas de trabajos obtenidas exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Estadísticas de trabajos obtenidas exitosamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "total_jobs": {
                                                    "type": "integer",
                                                    "example": 150
                                                },
                                                "active_jobs": {
                                                    "type": "integer",
                                                    "example": 120
                                                },
                                                "inactive_jobs": {
                                                    "type": "integer",
                                                    "example": 30
                                                },
                                                "jobs_by_type": {
                                                    "type": "object"
                                                },
                                                "jobs_by_location": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}